refactor: レビュー対応(生成物のビルド時取得化・型とモジュール構成の最小化)
- 日次画像と daily-art.json をリポジトリから除外し、ビルド時に取得する方式に変更
  (npm prebuild で fetch、daily-art.yml は deploy をディスパッチするだけに簡素化)
- MediaType を実際に使う 'interactive' のみに縮小し、GallerySection から
  ライトボックスと動画分岐を削除
- 未使用の lib/cv/contours を削除。lib/gl・lib/cv を作品スコープの
  lib/weathering/(engine.ts, gradientField.ts)に集約
- 作品ページの title/description は gallery.yaml を SSOT として参照
- tools から Computer Vision を削除
