+++ .gitignore
@@ -13,3 +13,7 @@ pnpm-debug.log*
 references/
 memo.md
 .serena/
+
+# generated at build time by scripts/fetch-daily-art.mjs
+/data/daily-art.json
+/public/images/gallery/daily/
