+++ components/GallerySection.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState } from 'react';
-import type { ReactNode, SyntheticEvent } from 'react';
+import type { SyntheticEvent } from 'react';
 import { useRouter } from 'next/navigation';
 import {
   Container,
@@ -11,13 +11,9 @@ import {
   CardMedia,
   CardContent,
   Chip,
-  Dialog,
-  DialogContent,
-  IconButton,
   Tabs,
   Tab,
 } from '@mui/material';
-import CloseIcon from '@mui/icons-material/Close';
 import type { GalleryWork } from '@/types/gallery';
 import { GalleryEmptyState } from './GalleryEmptyState';
 
@@ -29,7 +25,6 @@ type GallerySectionProps = {
 
 export function GallerySection({ works, showTitle = true, showCategories = true }: GallerySectionProps) {
   const router = useRouter();
-  const [selectedWork, setSelectedWork] = useState<GalleryWork | null>(null);
   const [selectedCategory, setSelectedCategory] = useState<string>('all');
 
   // 作品が0件の場合はEmptyStateを表示
@@ -55,53 +50,8 @@ export function GallerySection({ works, showTitle = true, showCategories = true 
       ? works
       : works.filter((work) => work.category === selectedCategory);
 
-  const handleWorkClick = (work: GalleryWork) => {
-    // Interactive works live on their own page instead of the lightbox.
-    if (work.mediaType === 'interactive') {
-      router.push(work.media);
-      return;
-    }
-    setSelectedWork(work);
-  };
-
-  const handleCloseDialog = () => {
-    setSelectedWork(null);
-  };
-
   const handleCategoryChange = (_event: SyntheticEvent, newValue: string) => {
     setSelectedCategory(newValue);
-  };
-
-  const renderMedia = (work: GalleryWork, isThumbnail = true): ReactNode => {
-    const src = isThumbnail ? work.thumbnail : work.media;
-
-    if (work.mediaType === 'video') {
-      return (
-        <video
-          src={src}
-          controls={!isThumbnail}
-          autoPlay={!isThumbnail}
-          loop
-          style={{
-            width: '100%',
-            height: isThumbnail ? 240 : 'auto',
-            objectFit: 'cover',
-          }}
-        />
-      );
-    }
-
-    return (
-      <CardMedia
-        component="img"
-        image={src}
-        alt={work.title}
-        sx={{
-          height: isThumbnail ? 240 : 'auto',
-          objectFit: 'cover',
-        }}
-      />
-    );
   };
 
   return (
@@ -157,15 +107,18 @@ export function GallerySection({ works, showTitle = true, showCategories = true 
                 boxShadow: 4,
               },
             }}
-            onClick={() => handleWorkClick(work)}
+            onClick={() => router.push(work.media)}
           >
-            {renderMedia(work, true)}
+            <CardMedia
+              component="img"
+              image={work.thumbnail}
+              alt={work.title}
+              sx={{ height: 240, objectFit: 'cover' }}
+            />
             <CardContent>
               <Typography variant="h6" component="h3" gutterBottom>
                 {work.title}
-                {work.mediaType === 'interactive' && (
-                  <Chip label="Interactive" size="small" sx={{ ml: 1, verticalAlign: 'middle' }} />
-                )}
+                <Chip label="Interactive" size="small" sx={{ ml: 1, verticalAlign: 'middle' }} />
               </Typography>
               <Typography variant="caption" color="text.secondary" display="block" sx={{ mb: 1 }}>
                 {new Date(work.date).toLocaleDateString('ja-JP', {
@@ -191,85 +144,6 @@ export function GallerySection({ works, showTitle = true, showCategories = true 
           </Card>
         ))}
       </Box>
-
-      {/* Lightbox Dialog */}
-      <Dialog
-        open={!!selectedWork}
-        onClose={handleCloseDialog}
-        maxWidth="lg"
-        fullWidth
-        slotProps={{
-          paper: {
-            sx: {
-              bgcolor: 'background.paper',
-            },
-          },
-        }}
-      >
-        <IconButton
-          onClick={handleCloseDialog}
-          sx={{
-            position: 'absolute',
-            right: 8,
-            top: 8,
-            color: 'text.primary',
-            zIndex: 1,
-            bgcolor: 'background.paper',
-            '&:hover': {
-              bgcolor: 'action.hover',
-            },
-          }}
-        >
-          <CloseIcon />
-        </IconButton>
-
-        {selectedWork && (
-          <DialogContent sx={{ p: 0 }}>
-            <Box sx={{ position: 'relative' }}>
-              {renderMedia(selectedWork, false)}
-            </Box>
-            <Box sx={{ p: 3 }}>
-              <Typography variant="h4" component="h2" gutterBottom>
-                {selectedWork.title}
-              </Typography>
-              <Typography variant="body2" color="text.secondary" sx={{ mb: 2 }}>
-                {new Date(selectedWork.date).toLocaleDateString('ja-JP', {
-                  year: 'numeric',
-                  month: 'long',
-                  day: 'numeric',
-                })}
-              </Typography>
-              {selectedWork.description && (
-                <Typography variant="body1" sx={{ mb: 2 }}>
-                  {selectedWork.description}
-                </Typography>
-              )}
-              {selectedWork.tools && selectedWork.tools.length > 0 && (
-                <Box>
-                  <Typography variant="subtitle2" color="text.secondary" gutterBottom>
-                    Tools Used
-                  </Typography>
-                  <Box sx={{ display: 'flex', gap: 1, flexWrap: 'wrap' }}>
-                    {selectedWork.tools.map((tool) => (
-                      <Box
-                        key={tool}
-                        sx={{
-                          px: 2,
-                          py: 0.5,
-                          bgcolor: 'action.selected',
-                          borderRadius: 1,
-                        }}
-                      >
-                        <Typography variant="body2">{tool}</Typography>
-                      </Box>
-                    ))}
-                  </Box>
-                </Box>
-              )}
-            </Box>
-          </DialogContent>
-        )}
-      </Dialog>
     </Container>
   );
 }
