--- data/daily-art.json
@@ -1,141 +1,0 @@
-{
-  "date": "2026-07-03",
-  "seed": 3365759791,
-  "thumbnail": "/images/gallery/daily/current-thumb.jpg",
-  "works": [
-    {
-      "image": "/images/gallery/daily/0.jpg",
-      "event": {
-        "year": 1935,
-        "text": "André Citroën, French engineer and businessman, founded the Citroën Company (born 1878)",
-        "pageTitle": "André Citroën",
-        "pageUrl": "https://en.wikipedia.org/wiki/Andr%C3%A9_Citro%C3%ABn"
-      },
-      "source": {
-        "fileTitle": "File:AndreCitroen.jpg",
-        "filePageUrl": "https://commons.wikimedia.org/wiki/File:AndreCitroen.jpg",
-        "license": "Public domain",
-        "artist": "Bain News Service"
-      }
-    },
-    {
-      "image": "/images/gallery/daily/1.jpg",
-      "event": {
-        "year": 1938,
-        "text": "Rhoda Scott, American soul and jazz Hammond organist",
-        "pageTitle": "Rhoda Scott",
-        "pageUrl": "https://en.wikipedia.org/wiki/Rhoda_Scott"
-      },
-      "source": {
-        "fileTitle": "File:PhotoRhodaScottpro.JPG",
-        "filePageUrl": "https://commons.wikimedia.org/wiki/File:PhotoRhodaScottpro.JPG",
-        "license": "Public domain",
-        "artist": "Concertparis"
-      }
-    },
-    {
-      "image": "/images/gallery/daily/2.jpg",
-      "event": {
-        "year": 1943,
-        "text": "Norman E. Thagard, American astronaut",
-        "pageTitle": "Norman Thagard",
-        "pageUrl": "https://en.wikipedia.org/wiki/Norman_Thagard"
-      },
-      "source": {
-        "fileTitle": "File:Thagard-ne.jpg",
-        "filePageUrl": "https://commons.wikimedia.org/wiki/File:Thagard-ne.jpg",
-        "license": "Public domain",
-        "artist": "NASA"
-      }
-    },
-    {
-      "image": "/images/gallery/daily/3.jpg",
-      "event": {
-        "year": 1866,
-        "text": "Austro-Prussian War is decided at the Battle of Königgrätz, enabling Prussia to exclude Austria from German affairs.",
-        "pageTitle": "Austro-Prussian War",
-        "pageUrl": "https://en.wikipedia.org/wiki/Austro-Prussian_War"
-      },
-      "source": {
-        "fileTitle": "File:Schlacht-bei-koeniggraetz-von-georg-bleibtreu.jpg",
-        "filePageUrl": "https://commons.wikimedia.org/wiki/File:Schlacht-bei-koeniggraetz-von-georg-bleibtreu.jpg",
-        "license": "Public domain",
-        "artist": "Georg Bleibtreu"
-      }
-    },
-    {
-      "image": "/images/gallery/daily/4.jpg",
-      "event": {
-        "year": 1728,
-        "text": "Robert Adam, Scottish-English architect, designed Culzean Castle (died 1792)",
-        "pageTitle": "Robert Adam",
-        "pageUrl": "https://en.wikipedia.org/wiki/Robert_Adam"
-      },
-      "source": {
-        "fileTitle": "File:Robert-adam.jpg",
-        "filePageUrl": "https://commons.wikimedia.org/wiki/File:Robert-adam.jpg",
-        "license": "Public domain",
-        "artist": "Attributed to George Willison"
-      }
-    },
-    {
-      "image": "/images/gallery/daily/5.jpg",
-      "event": {
-        "year": 1886,
-        "text": "The New-York Tribune becomes the first newspaper to use a linotype machine, eliminating typesetting by hand.",
-        "pageTitle": "New-York Tribune",
-        "pageUrl": "https://en.wikipedia.org/wiki/New-York_Tribune"
-      },
-      "source": {
-        "fileTitle": "File:Nytrib1864.jpg",
-        "filePageUrl": "https://commons.wikimedia.org/wiki/File:Nytrib1864.jpg",
-        "license": "Public domain",
-        "artist": "Unknown author"
-      }
-    },
-    {
-      "image": "/images/gallery/daily/6.jpg",
-      "event": {
-        "year": 1986,
-        "text": "Rudy Vallée, American singer, saxophonist, and actor (born 1901)",
-        "pageTitle": "Rudy Vallée",
-        "pageUrl": "https://en.wikipedia.org/wiki/Rudy_Vall%C3%A9e"
-      },
-      "source": {
-        "fileTitle": "File:Rudyvallee.jpg",
-        "filePageUrl": "https://commons.wikimedia.org/wiki/File:Rudyvallee.jpg",
-        "license": "Public domain",
-        "artist": "Bain News Service"
-      }
-    },
-    {
-      "image": "/images/gallery/daily/7.jpg",
-      "event": {
-        "year": 1938,
-        "text": "United States President Franklin D. Roosevelt dedicates the Eternal Light Peace Memorial and lights the eternal flame at Gettysburg Battlefield.",
-        "pageTitle": "Gettysburg Battlefield",
-        "pageUrl": "https://en.wikipedia.org/wiki/Gettysburg_Battlefield"
-      },
-      "source": {
-        "fileTitle": "File:Gettysburg-Open-Battlefield.svg",
-        "filePageUrl": "https://commons.wikimedia.org/wiki/File:Gettysburg-Open-Battlefield.svg",
-        "license": "Public domain",
-        "artist": "Veggies"
-      }
-    },
-    {
-      "image": "/images/gallery/daily/8.jpg",
-      "event": {
-        "text": "Christian feast day:\nJuly 3 (Eastern Orthodox liturgics)",
-        "pageTitle": "July 3 (Eastern Orthodox liturgics)",
-        "pageUrl": "https://en.wikipedia.org/wiki/July_3_(Eastern_Orthodox_liturgics)"
-      },
-      "source": {
-        "fileTitle": "File:East-Ortho-cross.svg",
-        "filePageUrl": "https://commons.wikimedia.org/wiki/File:East-Ortho-cross.svg",
-        "license": "Public domain",
-        "artist": "njk92"
-      }
-    }
-  ]
-}
