+++ lib/weathering/gradientField.ts
