+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
+    "prebuild": "npm run fetch-daily-art",
     "build": "next build --turbopack",
     "start": "next start",
     "lint": "eslint",
