+++ types/gallery.ts
@@ -1,10 +1,10 @@
-export type MediaType = 'image' | 'gif' | 'video' | 'interactive';
+export type MediaType = 'interactive';
 
 export type GalleryWork = {
   id: string;
   title: string;
   date: string;
-  media: string; // Path to media file (route path for interactive works)
+  media: string; // Route path to the work page
   thumbnail: string; // Path to thumbnail
   mediaType: MediaType;
   description?: string;
